requirements.txt
openai
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib